requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pylint==2.17.5
requests==2.31.0
Cerberus==1.3.5
gunicorn==21.2.0
